Rename module hit to scene
