+++ src/materials/dielectric.rs
@@ -1,5 +1,5 @@
 use crate::core::{Color, Vec3, Ray};
-use crate::hit::BasicHitRecord;
+use crate::scene::BasicHitRecord;
 use crate::util::rand;
 
 use super::material::{Material, ScatterRecord, schlick};
