+++ src/materials/lambertian.rs
@@ -1,6 +1,6 @@
 use crate::core::{Color, Ray};
 use crate::util::rand_unit_vector;
-use crate::hit::BasicHitRecord;
+use crate::scene::BasicHitRecord;
 
 use super::material::{Material, ScatterRecord};
